refactor: split tab-manager into component-registry, workspace-layout and sidebar-manager
