+++ src/components/file-viewer.js
@@ -6,6 +6,7 @@ import { contextMenu } from './context-menu.js';
 import { generateId } from '../utils/id.js';
 import { _el, setupInlineInput } from '../utils/dom.js';
 import { getCursorPosition, insertTab, parseWebviewUrl, SAVE_FLASH_MS, TAB_SPACES, EMPTY_MESSAGE, STATIC_MODES, MODE_CONFIG, ALL_STATIC_ELEMENTS, MODE_ACTIVATE, pinnedFiles } from '../utils/editor-helpers.js';
+import { registerComponent } from '../utils/component-registry.js';
 
 export class FileViewer {
   constructor(container, isActive) {
@@ -467,3 +468,5 @@ export class FileViewer {
     this._webviewEls.clear();
   }
 }
+
+registerComponent('FileViewer', FileViewer);
